+++ lib/ftpd/session.rb
@@ -4,6 +4,7 @@ module Ftpd
   class Session
 
     include Error
+    include ListPath
 
     def initialize(opts)
       @failed_login_delay = opts[:failed_login_delay]
@@ -201,8 +202,7 @@ module Ftpd
         ensure_logged_in
         ensure_file_system_supports :dir
         ensure_file_system_supports :file_info
-        path = argument
-        path ||= '.'
+        path = list_path(argument)
         path = File.expand_path(path, @name_prefix)
         transmit_file(list(path), 'A')
       end
@@ -212,8 +212,7 @@ module Ftpd
       close_data_server_socket_when_done do
         ensure_logged_in
         ensure_file_system_supports :dir
-        path = argument
-        path ||= '.'
+        path = list_path(argument)
         path = File.expand_path(path, @name_prefix)
         transmit_file(name_list(path), 'A')
       end